trace: cheri support for events
Update the event framework for cheri.

Generally, an event defines a struct with data that is collected for each
occurrence. An event is linked to one or more probe functions. All of
them are executed when the event occurs, they receive the collected data.
For each event, the kernel provides a default probe function that logs
the event data into the ringbuffer.

It's common for events to have pointers in their struct. They must be
converted to capabilities for cheri. The probe functions may want to
dereference such pointers.

The ringbuffer on a cheri system does not store capabilities. The
required alignment would reduce the buffer's capacity by a lot.

The basic idea for adapting trace events for cheri is to log a capability
as two separate fields: address and tag+metadata.

We keep the existing field name for the address field. Filter expressions
will continue to work as before.

Instead of using a separate field for the tag bit, we store it in one of
the reserved 0 bits of the metadata. Yet again, another field would waste
space for alignment.

Splitting the capabilities into the two fields has to be done while the
tracing is running. This is a rather simple operation.

Dissection of the metadata and bounds calculation are done when the
logged events are read from the ringbuffer. The existing printk format
for capabilities can be used.

The downside of this approach (and probably of any other alternative) is
that the event definitions have to be rewritten. We need new macros for
pointers. Most of these changes can be automated by a script.