cheri: gitlab-ci: Ignore clang-tidy messages on older kernels
The docker container includes one clang-tidy binary and is used for
all versions of the kernel. However, the list of ioctls changes in
incompatible ways between kernel versions and other changes might
require us to ignore some messages on older kernels.

Add a simple mechnism to achieve this. The lastest version of the
kernel should always have an empty ignore file.