cheri: Introduce purecap UAPI types
Define __kernel_uintptr_t and related types as a capability when in
purecap UAPI mode and use this to correctly define the in kernel
and UAPI types.

Here's a list of the UAPI types an their (new) sizes for reference:

TYPE		64-bit		64-bit		32-bit		32-bit
		cheri		non-cheri	cheri		non-cheri

__uptr		128		64		64		32
unsigned long	 64		64		32		32
__u64ptr	128		64		64		64
__u64		 64		64		64		64

There is a complication for cases where the kernel does not use
capabilities internally but the user space UAPI does. In this
case UAPI headers must use the capability types when used inside
the kernel even though __CHERI_PURE_CAPABIITY__ is not defined.
Define __ARCH_WANT_PURECAP in include/linux/types.h and check
for this in addition to __CHERI_PURE_CAPABILITY__ in the UAPI
headers.