ring-buffer: fix event data len for 8-byte-alignment
If 8-byte alignment is required for a ring buffer data event, array[0]
contains a length field and the data starts at array[1]. The same format
is used when the length cannot be encoded in type_len. The length field
includes the size of array[0] itself. This can be seen in
rb_calculate_event_length.

The calculation of the actual data length from the length field must
subtract sizeof(array[0]). This is done only for the case where type_len
is too small, not for forced 8-byte alignment. Fix this calculation.

Signed-off-by: Martin Kaiser <martin@kaiser.cx>