module: Reference symbol via pointer instead of unsigned long
Change the "struct kernel_symbol" to reference the actual symbol via a
pointer and not via unsigned long. This requires that we adjust the
__KSYMTAB macro to use .chericap when possible and declare .size for
the two strings of the symbol.

As bonus, once we reference the symbol via a pointer, we get the size
of the symbol on architectures with hardware capabilities. This will
be later relevant, once we enable dynamic module loading, where we
need the information about the symbol's size to derive the
capabilities when relocations are processed.

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>