module: Introduce shdr_addr() getter
The kernel, and in particular its module subsystem, often derives a
pointer from a sh_addr member of Elf_Shdr. However, this member is not
a pointer and its type can not be easily changed to a pointer without
creating a mismatch with the on-disk format of an ELF.

On platforms with architectural capabilities this means that we need
to use the metadata in the ELF section header to derive a capability
that is able to access the data of the section.

The new shdr_addr() function does exactly that.

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>