riscv: module: Support kernel-module loading with CHERI
To support kernel modules with CHERI we need to derive capabilities
for each symbol referenced by a location.

One central function is derive_capability() which is to materialize
capabilities. Depending on the relocation type, those capabilities are
used to populate the captable. In this case, we need to know the
address of the capability, its size and its permissions (basically RO,
RW, or RX). Obtaining the address is easy, it's already handed to us
as part of the standard relocation machinery of the kernel. Obtaining
the size is done via the Elf_Sym* provided to the relocation
handling. Determining the permissions is a bit more complicated. We
need to see if the address is part of one of the know sections of the
kernel. This includes the memory of other modules and per-cpu data.

We use two captables: .captable and .init.captable.

The first, .captable, holds all capbilities required by relocations
in non-init segments. The second, .init.captable holds all capability
required by relocations in *init* segments. The .init.captable will be
deallocated together with the memory of the init segments.

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>