riscv: dts: codasip: Change to Hobgoblin2
Hobgoblin2 changes the memory map to free up memory space for upcoming
PCIe support. PCI support needs large chunks of contiguous address
space, the older memory map was somewhat profligate in using address
space. This is compounded by the Xilinx PCIe controller having some
restrictions about alignment of regions.

The DRAM and SRAM blocks stay where they were, but the
SRAM blocked is doubled in size to 1M. The peripherals are moved from
0x60000000 to 0x10000000. The ROM is shifted from 0x10000000 to
0x28000000. Ideally you would have the ROM at a lower address, but one
of the goals was not to move the location of SRAM/DRAM to have less of
an impact on software. Other minor changes are that the Ethernet irqs
are moved to below 16 to make it possible to use it on platforms with
lower numbers of interrupts.

The result of these changes is that the whole area from 0x40000000 to
0x7fffffff is cleared out for use by PCI express.

Signed-off-by: David McKay <david.mckay@codasip.com>