riscv: dts: codasip: Get MAC address from EEPROMs
This patch allows the kernel and bootloaders to pull the MAC addresses
out of the various EEPROMs on the board.

The VCU118 has board information in the motherboard EEPROM
connected via the mux chips. This is nominated as ethernet0.

The Ethernet on the FMCLite is held in the internal area of the VITA 57
data on the FMCLite EEPROM. This is ethernet1.

Signed-off-by: David McKay <david.mckay@codasip.com>