riscv: cheri: resize dyn ftrace nop region
For dynamic ftrace, the compiler generates some additional space around
the start of each traceable function.

---- start of the compiler-generated "nop area"
    <pointer to a struct ftrace_ops>
---- start of the function
    auipc that loads ftrace trampoline's address
    nop or jalr
    <actual function code>

The ftrace_ops pointer is used by the trampoline code to access the
struct and read e.g. the trampoline's destination. For cheri, this
pointer must be a capability.

This patch changes the compiler flags to add 16 bytes of space before the
start of the function. Storing a capability in this area requires it to
be 16-byte aligned. We align the start of the function to 16 bytes to get
the correct alignment for the preceding pointer.