riscv: dts: codasip: PCI Express support
This adds support for PCIe on the VCU118 board using the Opsero
FPGADrive FMC card. Two controllers are added.

The xilinx driver is configured in to use the interrupt decode mode,
rather than the legacy FIFO mode, as the legacy mode could not handle
the interrupt rate with such a relatively slow CPU. However, I suspect
the driver has some problems in legacy mode anyway.

Signed-off-by: David McKay <david.mckay@codasip.com>