riscv: ftrace: update mcount-dyn.S for cheri
Update the dynamic ftrace assembler code in mcount-dyn.S for cheri.

Most changes to registers and instructions are straightforward.

The trampoline for dynamic ftrace is ftrace_caller. It receives the
return address in t0. There is a struct ftrace_ops in front of the
calling function.

The non-cheri code uses t0 to access the struct ftrace_ops. For cheri, t0
is sealed and does not grant read access. We have to fabricate a new
capability with t0's address and pcc's bounds and permissions.

The offset between t0 (or the fabricated capability) and struct
ftrace_ops is different for cheri vs. non-cheri. Use asm-offsets to
generate a define for it.