ftrace: cherify the code that patches ftrace call-sites
For dynamic ftrace on risc-v, the compiler adds some nop instructions
around the start of each function that can be traced. It also creates a
list of all those nop areas. Their start addresses are available in the
__mcount_loc section of the kernel image.

The initialisation of the kernel populates each nop area as follows

---- this address is stored by the compiler in __mcount_loc
    <pointer to a struct ftrace_ops>
---- start of the function
    auipc that loads ftrace trampoline's address
    nop
    <actual function code>

When tracing is enabled for a function, the nop is replaced with a jalr.
Calling the function will then jump to the trampoline.

This patch is the first step of adapting this mechanism for cheri.

When compiling for cheri, the compiler still inserts addresses into
__mcount_loc, it does not create any capability metadata. The kernel
parses these addresses and creates a struct dyn_ftrace for each one.

Before patching any code as shown above, we have to create a capability.
This could be done when populating the struct dyn_ftrace entries at
startup or when code is patched.

There's one struct dyn_ftrace for each tracable function. Storing a
capability in there would increase the memory footprint significantly and
would require updating all code that handles struct dyn_ftrace.

Fabricating a capability when the code is patched avoids these downsides
and allows limiting the capability bounds to the area that is actually
patched.