cheri: ptrace: Only report addresses for syscall info
We do not want to give user space tagged capabilities in
struct ptrace_syscall_info. Thus refrain from using
capability data at all. We might later revisit this and
come up with an extension that can be used to report capability
data including tags.

This fixes strace which depends on the format.