riscv: dts: codasip: Remove reg-offset from 16550 UART
The Xilinx 16550 uart implementation has a built in offset of 0x1000 for
some reason. This is why the reg-offset property is used. However, it
turns out that BSD at least does not support this property, so remove it
from the description. As a side note, the Xilinx UART actually works
fine if you do not have the offset present, it decodes at the base
address anyway!

Signed-off-by: David McKay <david.mckay@codasip.com>