riscv: ftrace: update the rest of mcount-dyn.S for cheri
mcount-dyn.S has some code that's run only if dynamic ftrace is used
without call ops. For risc-v, call ops are always enabled unless the
compiler uses cfi (both call ops and cfi modify the start of traceable
functions).

This patch updates the code for dyn ftrace, no call ops to work with
cheri. It was tested by fiddling with the kernel config to use this
combination regardless of cfi.