ftrace: make direct_call a capability
Make direct_call in struct ftrace_ops a capability. It is written to a
register and used as a destination for a jump instruction.

Update all functions that generate or relay values for direct_call.

Don't convert "direct" in struct ftrace_func_entry. An address is
sufficient for hash calculation and symbol lookup.