cheri: trace: make trace_printk work for cheri
trace_printk is a printf-style function that writes into the tracing
ringbuffer. There's optimisations to avoid processing the format
specifiers, e.g. if there's only one argument or if none of the arguments
contain format specifiers.

For these optimisations, we have a constant string in .rodata and a
pointer to it in __trace_printk_fmt. The pointer will be stored in the
ringbuffer. Pointers/capabilities in the ringbuffer are not supported for
cheri. Aligning ringbuffer entries to capability size would waste too
much space.

Drop the optimisations for cheri. Overwrite the check for constant
strings. Given that trace_printk is rarely used, this simple fix should
be ok for now.