kunit: Cast priv pointer to its actual type test_dev_action()
Avoid an cast in test_dev_action() by casting the priv pointer to its
actual type (long*). The previous cast was also wrong because original
type of the void* is long* and not void**.

As an additional benefit, this fixes an issue where a CHERI-enabled
compiler may synthesize a null capability for the priv access, leading
to a CHERI fault.

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>