cheri: mmap: Fix reservation check for MAP_FIXED
When doing a MAP_FIXED mapping within an existing reservation,
the existing reservation will be suitably aligned to create bounds
for the new mmap that do not exceed the reservation.

Thus it is sufficient to check the unaligned range against the
inner bounds of the existing reservation. If we don't do this
some legitimate MAP_FIXED mmaps may fail. An alternative would
be to check the aligned range against the outer bounds of the
existing reservation.