riscv: Support custom extensions for kernel build
Add a string-typed config option that is appended to march during
kernel compile. This enabled kernels compiled for a particular
target machine with known extensions.

Use the CHERI prime list of extension in the default codasip-*730
configs.

While there enable the really useful earlycon on SBI config option.
Enable it at runtime with "earlycon=sbi" on the command line.