cheri: riscv: Use ZABHLRSC extension for atomics
The riscv base ISA does not have atomics (AMO or lr/sc) that
operate on objects smaller than 32-bit. The fallback method
for this in standard riscv is to operate on a 32-bit word
in memory and use masking to achieve the desired effect.

While this seems to work for CHERI in practice it is technically
incorrect as we might access a large memory chunk than authorised
by the pointer capability.

The CHERI spec realizes that this is a problem and provides the
ZABHLRSC extension which defines the missing {lr,sc}.{b,h}
instructions for byte and half word memory accesses.

Use these instrucitons in cmpxchg.h if they are present to replace
the masked xchg and cmpxchg variants.

Detection of toolchain support is a bit tricky:
As of llvm-17 the instructions are recognized if the PCUABI
is used but the extension name itself is not recognized. Thus
detect if the toolchain supports the extension name and if it
does add it the the ISA string.

For runtime detection simply assume that the instructions are
available iff we are compiling a CHERI kernel. Do not depend
on the extension being declard in the device tree because as of
this writing the auto generated device trees in qemu do not declare
it and legacy device trees might not know that this is an
extension, now.

Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>