cheri: codasip: Update codasip configs
Enable support for graphics, input devices and the bootup logo.
This allows X to run when booting on the qemu "virt" platform
with virtio GPU and input devices.

Additionally, remove the fixed ISA string from the non-cheri
kernel config for a730. Reportedly this breaks boot on qemu and
most likely at least zish4add is not available on that
configuration anyway. If we really need on optimized version
someone with access to the real hardware will have to provide
a correct ISA string.