riscv: dts: codasip: Correct GPIO buttons
The GPIOs assigned to the buttons are wrong for later builds of the
FPGA. The reason for this is that the VCU118 only has 4 switches,
whereas the genesysII board had 8, and the design used to just have
unconnected GPIOs. This has been changed and the unused GPIOs removed,
so the buttons have moved down.

Signed-off-by: David McKay <david.mckay@codasip.com>