trace: cheri: update syscall_enter events for cheri
Add cheri support for the syscall_enter tracepoints and events which are
created for each syscall.

Reduce the syscall arguments to addresses, the ringbuffer doesn't support
capabilities. Don't modify syscall_get_arguments, this function is used
outside of the tracing code. Instead, reduce the registers when they're
copied to the ringbuffer.