cheri: io_uring: Cherify big CQEs
If an io_uring is set up with IORING_SETUP_CQE32, the CQE size is
increased to fit two 8-byte values that are commonly accessed via
'extra1' and 'extra2' in the inferred struct 'io_big_cqe'. With CHERI,
the CQE size is increased to fit two 16-byte values if using
IORING_SETUP_CQE32. Thus, change the type of the CQE member as well as
'extra1' and 'extra2' to a type that fits a capability, but is at least
64-bit big, i.e. the type __u64ptr.

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>