trace: events: cheri support for pointers to strings
The vanilla kernel supports events that contain pointers to strings.

When the event occurs, a pointer to a string is logged as part of the
event. When the logged data is read and parsed later, the pointer is
dereferenced to get the string.

By the time the pointer is dereferenced, it might no longer be valid or
the data it points to might have changed since the event was logged. Or
the parsing of the logged event might be done by libtraceevent in
userspace, where the kernel pointer can't be dereferenced.

To address all of these problems, the pointers that are used in events
must point to compile-time constant data and must be marked as
tracepoint_string. The kernel will put all those pointers in the
__tracepoint_str section. userspace applications can read all addresses
and corresponding data from /sys/kernel/tracing/printk_formats

This patch adapts the handling of string pointers for cheri.

On cheri, the __tracepoint_str section contains an array of capabilities
to access the constant strings of the events.

A pointer to a string is logged as address, tag+metadata like any other
capability. When the pointer is dereferenced on a cheri system, we use
its address to search its capability in the __tracepoint_str section. The
section has only a few entries, a linear search is sufficient for now.