module: Invalidate mod->init after module initialized
The memory backing the module initialization code is deallocated after
it runs, leaving mod->init as a dangling pointer (or capability, in
CHERI terms). When using CHERI architectural capabilities, we want to
prevent the mod->init capability from being re-used to potentially
access re-purposed memory. Therefore, we override the capability with
an invalid one after the initialization codes completes.

Note that we can not set mod->init to NULL, because it is used later
in delete_module() to check if the module had an init function.

Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>