cheri: arm64: compat: Use correct syscall handlers
When doing 64-bit compat for cheri (as opposed to 32-bit
compat) several system calls do not need a compat version
because they only deal with 32-bit vs. 64-bit unsigned long.
They still need compat entry code, though. If this is the case
the handle that the compat entry code jumps to is re-defined
to point to the 64-bit syscall.

After rebasing to v6.16 there is a 32-bit syscall table and a 64-bit
one and for the 64-bit compat code we use compat entries in the
64-bit table. All the time related syscalls do not have/need a
compat version in the 64-bit table anymore and thus there is no
need to re-define them.

Actually call the compat version of the syscall from the compat
entry code when using the 64-bit table and remove the defines
that are no longer necessary as a result of this.