cheri: firewire: Make closure fields in UAPI capability sized
The firewire UAPI uses fields called "closure" all over its UAPI.
These fields are mostly kernel opaque value that might be echoed
back to user space on completion of some request.

For the PCUABI we must keep the tag on any capability that is passed
in such a field. Thus make the fields capability sized and make sure
that tags are not stripped in the kernel.