cheri: Add auto generated compat headers and tools
There is a clang-tidy "check" that annotates UAPI structures
with "warnings". These warnings can be parsed and used to generate
compat32 and compat64 versions of the same structs.

The definition of a compat64 version is that the compat64 version
of the structure when used in a purecap 64-bit kernel has the
exact same layout that the same structure would have when compiled
for a non-cheri 64-bit kernel. Similar definitions apply to the
compat32 version.

This change adds the necessary script to generate the compat headers
and the generated compat64 headers.

Additionally, some files are added to the compat/ subdirectory that
can be used to check the offsets of the fields in the compat structures
and compare them to the same result for a non-purecap kernel.