cheri: Add perl script to generate compat headers
There is a clang-tidy "check" that annotates UAPI structures
with "warnings". These warnings can be parsed and used to generate
compat32 and compat64 versions of the same structs.

The generated compat64 version of a struct when used in a purecap
64-bit kernel has the exact same layout that the same structure
would have when compiled for a non-cheri 64-bit kernel.

This change adds the necessary script to generate the compat headers.

Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>