cheri: Add file to veriy auto-generated compat64 headers
Generate a list of compat64 includes and a file that contains
all structs and fields of compat64 structures. This can be used
to generate struct offsets similar to what asm-offsets.c does.

Generating these offsets once on a CHERI kernel and once on an
upstream vanilla kernel allows us to verify that the layout of the
auto generated compat64 structures matches the native 64-bit layout.

Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>