cheri: Update and check generated compat headers
The generated compat headers produce include files that need
manual editing because the structure definitions contain preprocessor
directives that cannot be handled by the automatic conversion.
Read through these files and adjust the compat struct definitions
as necessary.

Add a file compat/x64.c which when compiled to compat/x64.s will have
defines for offsets sizes and alignments of all fields in all generated
compat structures. Add the same file that is generated from a non-cheri
riscv64 compile as a reference. This can be used to verify that the
generated compat structure actually have the same layout as on a
non-cheri compile.