trace: syscalls: simplify sys_enter's print fmt
The sys_enter event contains an array of the raw syscall arguments. To
print an event, each array element is converted to a fake pointer, which
is then printed as a hex number.

Replicating this printing in userspace would require large changes to
libtraceevent. The combination of a custom function (__c_fakep) and array
indexing is unnecessarily complex.

This patch simpilifies the print format such that libtraceevent can
handle it without modifications. We print the array elements as unsigned
hex values rather than pointers. The __c_fakep call can then be removed
and libtraceevent's existing code handles the array elements.

The type of the array elements is updated to __u64, this is what __c_ua
returns. %llx is the correct format specifier for __u64.