io_uring: compat64: Fix sync cancel comparision for compat64
To cancel an IO task from the uring's work queue, user space provides in
'addr' the identifier that matches the user data of the to be cancelled
queue entry. In standard 64-bit ABI, user data is an unmodified 64-bit
unsigned integer. A pure cap kernel must downgrade 'addr' to a 64-bit
value as its compat capability would no longer match the 64-bit user
data in a full capability comparision.

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>