riscv: csum: Avoid over-read if CHERI is enabled
The RISC-V checksum implementation performed over-reads, which could
trigger CHERI faults due to out-of-bounds accesses. To avoid this, we
replace the over-reads with memcpy() when building a CHERI kernel.

As an additional benefit, we can drop __no_sanitize_address from the
CHERI variant of do_csum_common().

Thanks to Christian for the feedback that significantly improved this
solution.

Thanks-to: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>
Signed-off-by: Florian Schmaus <florian.schmaus@codasip.com>