cheri: io_uring: Fix handling of multi slot CQEs
The CQE ring can contain elements that span two slots in the ring.
There are two cases where this can happen:
- A traditional big CQE that includes up to two additional data
  items (__u64 or __u64ptr on CHERI)
- A CQE with payload data that does not follow this scheme. Instead
  data format of the second CQE depends on the type of the entry.

For the second case make sure that the caller provides the format
of the second CQE in compat64 format if this is required. Only the
first CQE is automatically converted to compat64 format.

To ensure that future instances of this are detected use a separate
parameter for the additional data. Its size must not be larger than
the compat64 version of an io_uring_cqe.

Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>