cheri: Introduce new UAPI types __u64ptr and __ulptr
With CHERI pointers are twice the size of an unsigned long.
In order to be able to use such a pointer in the UAPI, new
UAPI types are necessary. Additionally, pointers in CHERI
must be size aligned. This is enforced by the compiler but
it can conflict with forced packing of structs.

Introduce new UAPI types and macros that can be used to handle
this. In particular the new types are:
- __ulptr (replaced unsigned long): An integer with the natural
  size and alignment of a pointer.
- __u64ptr (replaces __u64): An integer with at least 64-bit
  that can hold a pointer. Alignment is also that of a pointer.
- __aligned_u64ptr (replaces __aligned_u64): The same as __u64ptr
  but always size aligned.
- The corresponding signed types are __slptr, __s64ptr and
  __aligned_s64ptr.
- __ptraddr_t: An integer that can hold the address part of
  a pointer but not a full capability on CHERI.
- __ptraddr64_t: A 64-bit integer that can hold the address part of
  a pointer.

Additionally, we introduce two macros that are used to handle
alignment and packed structures:
- __PACKED_IF_NOT_CHERI (replaces __atrribute__((packed))):
  The structure is packed only if the CHERI UAPI is not in use.
- __CHERI_POINTER_ALIGN: On CHERI systems this enforces an alignment
  of __SIZEOF_POINTER__ on a structure. Otherwise it is a NOP.

Signed-off-by: Christian A. Ehrhardt <christian.ehrhardt@codasip.com>