kernel/signal: compat64: Fix RT/timer signal copy_to/from_user
The compat32/64 conversion functions copy for RT signals the 32-bit
member 'sival_int' to/from below's union member:
	typedef union compat_sigval {
		compat_int_t	sival_int;
		compat_uptr_t	sival_ptr;
	} compat_sigval_t;

In case of compat64, 'sival_ptr' is 64-bit, so half of the bits are lost
if 'sival_int' is copied. Fix this by copying the larger union member
'sival_ptr'.

Do the same for timer signals, which are affected by the same problem.

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>